switch to discordjs12
